requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
astroid==2.2.5
autopep8==1.4.4
aws-shell==0.2.1
awscli==1.16.135
backcall==0.1.0
bashplotlib==0.6.5
beautifulsoup4==4.7.1
blessings==1.7
boto3==1.9.125
botocore==1.12.125
bs4==0.0.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
colorama==0.3.9
configobj==5.0.6
cryptography==2.7
curtsies==0.3.0
decorator==4.4.0
dj-database-url==0.5.0
Django==2.2.5
django-cors-headers==3.0.2
django-extensions==2.1.6
django-heroku==0.3.1
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-rest-passwordreset==1.0.0
djangorestframework==3.10.3
docutils==0.14
emoji==0.5.2
entrypoints==0.3
flake8==3.7.7
geographiclib==1.49
geopy==1.20.0
gunicorn==19.7.1
idna==2.8
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
isort==4.3.17
jedi==0.13.3
jmespath==0.9.4
kafka-python==1.4.6
lazy-object-proxy==1.3.1
mccabe==0.6.1
model-mommy==1.6.0
numpy==1.17.2
pandas==0.25.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prompt-toolkit==2.0.9
psycopg2==2.8.3
ptyprocess==0.6.0
pyasn1==0.4.5
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.3.1
pylint==2.3.1
python-dateutil==2.8.0
pytz==2019.2
PyYAML==3.13
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
tornado==6.0.3
traitlets==4.3.2
typed-ast==1.3.4
typing==3.6.6
urllib3==1.25.3
wcwidth==0.1.7
whitenoise==3.3.1
wrapt==1.11.1
